+++ test/server/Persistify.Management.Tests.Bench/Number/IntervalTreeNumberManagerBenchmark.cs
@@ -1,7 +1,6 @@
 ﻿using BenchmarkDotNet.Attributes;
 using Persistify.Management.Number.Manager;
 using Persistify.Management.Score;
-using Persistify.Protos.Documents;
 using Persistify.Protos.Documents.Shared;
 using Persistify.TestHelpers.Documents;
 using NumberQuery = Persistify.Management.Number.Search.NumberQuery;
@@ -11,9 +10,9 @@ namespace Persistify.Management.Tests.Bench.Number;
 [MemoryDiagnoser]
 public class IntervalTreeNumberManagerBenchmark
 {
-    private const uint TextFields = 3;
-    private const uint NumberFields = 3;
-    private const uint BoolFields = 3;
+    private const int TextFields = 3;
+    private const int NumberFields = 3;
+    private const int BoolFields = 3;
     private INumberManager _100KNumberManager = null!;
     private INumberManager _10KNumberManager = null!;
     private INumberManager _1KNumberManager = null!;
@@ -42,7 +41,7 @@ public class IntervalTreeNumberManagerBenchmark
         for (var i = 0; i < itemsCount; i++)
         {
             numberManager.Add("template", documentGenerator.GenerateDocument(TextFields, NumberFields, BoolFields),
-                (ulong)i);
+                i);
         }
 
         return numberManager;
